src/test/java/exercises/advanced/LambdaTest.java
package exercises.advanced;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.io.PrintStream;

import static java.lang.System.in;
import static java.lang.System.out;
import static java.lang.System.setIn;
import static java.lang.System.setOut;
import static org.assertj.core.api.Assertions.assertThat;

public class LambdaTest {
    private static final String INPUT = "5\n" +
            "1 4\n" +
            "2 5\n" +
            "3 898\n" +
            "1 3\n" +
            "2 12\n";

    private static final String OUTPUT = "EVEN\n" +
            "PRIME\n" +
            "PALINDROME\n" +
            "ODD\n" +
            "COMPOSITE\n";

    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
    private final ByteArrayInputStream inContent = new ByteArrayInputStream(INPUT.getBytes());

    @Before
    public void setUp() {
        setOut(new PrintStream(outContent));
        setIn(inContent);
    }

    @Test
    public void shouldSolveForFirstScenario() throws IOException {
        Lambda exercise = new Lambda();

        exercise.solve();

        assertThat(outContent.toString()).isEqualTo(OUTPUT);
    }

    @After
    public void tearDown() {
        setOut(out);
        setIn(in);
    }
}